pyproject.toml
[tool.poetry]
name = "s3stasher"
version = "0.0.2"
description = "A python package for working with objects in AWS S3 as if they were local files, including cache management and offline usage."
authors = ["Ben Siranosian <bsiranosian@gmail.com>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/bsiranosian/s3stasher"

[tool.poetry.dependencies]
python = "^3.10"
boto3 = "^1.36.21"
pytest = "^8.3.4"
pytest-cov = "^6.0.0"
pytz = "^2025.1"
python-dotenv = "^1.0.1"
tqdm = "^4.67.1"
tzlocal = "^5.3"
gitpython = "^3.1.44"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
